Move Ember versions file
